ebuild variable process spawner
A little messy, but essentially this introduces some new functions
which can spawn a persistent bash process, then pipe your ebuilds that
need to be sourced in, and it'll spit back out the sourced ebuilds variables to
stdout.

This is much faster than posix_spawning bash a trillion times, as we
can just have one (or several) bash process which will take in and process our ebuilds without worrying about kernel exec speeds.

Signed-off-by: Swagtoy <me@ow.swag.toys>